requirements.txt
pythonnet
PyQt6
configparser
Pillow
auto-py-to-exe
winregistry